Separate the implementation of Bench themes from the concept of a base theme. Bench light colors still serve as the base theme.
