+++ theme-list/themes/bench-light.js
@@ -1,0 +1,5 @@
+import { baseTheme } from '../../theme/src/theme';
+
+export const BenchLight = {
+  ...baseTheme,
+};
