+++ theme-panel/ThemePanel.tsx
@@ -4,16 +4,16 @@ import { ActiveTab, Tab, TabBar, ThemePanelContainer } from './ThemePanel.styles
 import { themeList } from '../theme-list/themeList';
 
 const ThemePanel: FC = ({ children }): ReactElement => {
-  const { theme, setTheme } = useTheme();
+  const { theme: currentTheme, setTheme } = useTheme();
 
   return (
     <ThemePanelContainer>
       <TabBar>
-        {Object.keys(themeList).map((key) =>
-          theme === themeList[key] ? (
-            <ActiveTab>{key}</ActiveTab>
+        {Object.entries(themeList).map(([themeName, theme]) =>
+          currentTheme === theme ? (
+            <ActiveTab>{themeName}</ActiveTab>
           ) : (
-            <Tab onClick={() => setTheme(themeList[key])}>{key}</Tab>
+            <Tab onClick={() => setTheme(theme)}>{themeName}</Tab>
           )
         )}
       </TabBar>
