+++ theme/doc/borderWidths.mdx
@@ -1,8 +1,8 @@
 import { mdx } from '@mdx-js/react';
 import { MdxLayout } from '~/layout';
-import { systemUILightTheme } from '../src/theme';
+import { baseTheme } from '../src/theme';
 import BorderWidthsShowcase from './components/border-widths-showcase/BorderWidthsShowcase';
 export default MdxLayout;
 
 # Border Widths
-<BorderWidthsShowcase theme={systemUILightTheme}/>
+<BorderWidthsShowcase theme={baseTheme}/>
