+++ theme/doc/fontWeight.mdx
@@ -1,9 +1,9 @@
 import { mdx } from '@mdx-js/react';
 import { SystemUIShowcases } from '@divriots/dockit-react/systemui-showcases';
 import { MdxLayout } from '~/layout';
-import { systemUILightTheme } from '../src/theme';
+import { baseTheme } from '../src/theme';
 export default MdxLayout;
 
 # Font Weight
 
-<SystemUIShowcases theme={systemUILightTheme} showcaseKey="fontWeight" />
+<SystemUIShowcases theme={baseTheme} showcaseKey="fontWeight" />
