+++ theme/doc/fonts.mdx
@@ -1,9 +1,9 @@
 import { mdx } from '@mdx-js/react';
 import { MdxLayout } from '~/layout';
 import FontsShowcase from './components/fonts-showcase/FontsShowcase';
-import { systemUILightTheme } from '../src/theme';
+import { baseTheme } from '../src/theme';
 export default MdxLayout;
 
 # FontsShowcase
 
-<FontsShowcase theme={systemUILightTheme}/>
+<FontsShowcase theme={baseTheme}/>
