+++ theme/doc/lineHeight.mdx
@@ -1,9 +1,9 @@
 import { mdx } from '@mdx-js/react';
 import { SystemUIShowcases } from '@divriots/dockit-react/systemui-showcases';
 import { MdxLayout } from '~/layout';
-import { systemUILightTheme } from '../src/theme';
+import { baseTheme } from '../src/theme';
 export default MdxLayout;
 
 # Line Height
 
-<SystemUIShowcases theme={systemUILightTheme} showcaseKey="lineHeight" />
+<SystemUIShowcases theme={baseTheme} showcaseKey="lineHeight" />
