+++ theme/doc/radii.mdx
@@ -1,13 +1,13 @@
 import { mdx } from '@mdx-js/react';
 import { SystemUIShowcases } from '@divriots/dockit-react/systemui-showcases';
 import { MdxLayout } from '~/layout';
-import { systemUILightTheme } from '../src/theme';
+import { baseTheme } from '../src/theme';
 export default MdxLayout;
 
 # Radii
 
 <SystemUIShowcases
-  theme={systemUILightTheme}
+  theme={baseTheme}
   showcaseKey="borderRadius"
   componentProps={{
     width: '8rem',
