+++ theme/src/variables.ts
@@ -1,0 +1,72 @@
+export const primary = {
+  '50': '#fce9c6',
+  '100': '#f9d38e',
+  '200': '#f0bd6b',
+  '300': '#e5a139',
+  '400': '#d17f13',
+  '500': '#b55e02',
+  '600': '#874903',
+  '700': '#633305',
+  '800': '#472300',
+  '900': '#331800'
+};
+
+export const secondary = {
+  '0': '#ffffff',
+  '20': '#f7f8fa',
+  '40': '#f0f1f4',
+  '60': '#ecedf0',
+  '80': '#e4e6e9',
+  '100': '#dcdfe6',
+  '200': '#cacdd2',
+  '300': '#abaeb5',
+  '400': '#8e9198',
+  '500': '#737780',
+  '600': '#565960',
+  '700': '#484b50',
+  '800': '#2d3036',
+  '820': '#202328',
+  '840': '#191b1e',
+  '860': '#121316',
+  '880': '#0a0b0c',
+  '900': '#000000'
+};
+
+export const success = {
+  '0': '#ebfcf8',
+  '100': '#b6f0e6',
+  '200': '#64d9c7',
+  '300': '#11bdaf',
+  '400': '#06a195',
+  '500': '#00877e',
+  '600': '#0b6660',
+  '700': '#104a46',
+  '800': '#013632',
+  '900': '#012421'
+};
+
+export const critical = {
+  '50': '#faefed',
+  '100': '#ffd9d6',
+  '200': '#ff9e99',
+  '300': '#fc6d62',
+  '400': '#f24d44',
+  '500': '#d6372b',
+  '600': '#ba1b13',
+  '700': '#85140c',
+  '800': '#520c02',
+  '900': '#3b0700'
+};
+
+export const highlight = {
+  '50': '#F3F6FB',
+  '100': '#dde7f8',
+  '200': '#b4caef',
+  '300': '#8aade5',
+  '400': '#5190f6',
+  '500': '#256af4',
+  '600': '#265bb0',
+  '700': '#1d4586',
+  '800': '#14305d',
+  '900': '#0b1a32'
+}
