+++ theme/stories/index.stories.tsx
@@ -1,5 +1,1 @@
-import { systemUIDarkTheme, systemUILightTheme } from '../src/theme';
-
-export const lightTheme = () => (systemUILightTheme);
-
-export const darkTheme = () => (systemUIDarkTheme);
+export const baseTheme = () => baseTheme;
